requirements.txt
pytube
youtube-transcript-api
moviepy
openai-whisper
transformers
streamlit
